+++ Kavita.Models/DTOs/UpdateChapterDto.cs
@@ -103,6 +103,6 @@ public sealed record UpdateChapterDto : IUpdateExternalMetadataIds
     public int? HardcoverId { get; set; }
     public long? MetronId { get; set; }
     public string? ComicVineId { get; set; }
-    public long? MangaBakaId { get; set; }
+    public int? MangaBakaId { get; set; }
     public int? CbrId { get; set; }
 }
