+++ Kavita.Models/Entities/Series.cs
@@ -137,7 +137,7 @@ public class Series : IEntityDate, IHasReadTimeEstimate, IHasCoverImage, IHasMet
     public bool IsStandAlone { get; set; }
     public long MetronId { get; set; }
     public string ComicVineId { get; set; }
-    public long MangaBakaId { get; set; } // TODO: Migrate this back to an int
+    public int MangaBakaId { get; set; }
     public string MangaBakaEditionId { get; set; }
     public int CbrId { get; set; }
     #endregion
