+++ Kavita.Models/Entities/Volume.cs
@@ -61,7 +61,7 @@ public class Volume : IEntityDate, IHasReadTimeEstimate, IHasCoverImage, IHasMet
     public int HardcoverId { get; set; }
     public long MetronId { get; set; }
     public string ComicVineId { get; set; }
-    public long MangaBakaId { get; set; }
+    public int MangaBakaId { get; set; }
     public int CbrId { get; set; }
 
     #endregion
